+++ src/forest5/backtest/grid.py
@@ -78,7 +78,7 @@ def run_grid(
     rsi_oversold: int = 30,
     rsi_overbought: int = 70,
     time_model: Path | None = None,
-    min_confluence: int = 1,
+    min_confluence: float = 1.0,
     n_jobs: int = 1,
     cache_dir: str = ".cache/forest5-grid",
     debug_dir: Path | None = None,
@@ -147,7 +147,7 @@ def run_grid(
         )
         settings.time.model.enabled = bool(time_model)
         settings.time.model.path = time_model
-        settings.time.fusion_min_confluence = int(min_confluence)
+        settings.time.fusion_min_confluence = float(min_confluence)
         res = run_backtest(df, settings)
         end, mdd, cagr = _compute_metrics(res.equity_curve)
         return GridResult(
